moved only function used in conf.py to utilities, deleted conf.py
