+++ gfail/utilities.py
@@ -2216,3 +2216,68 @@ def getFileType(filename):
     except BaseException:
         pass
     return "unknown"
+
+
+def correct_config_filepaths(input_path, config):
+    """
+    Takes an input filepath name and pre-pends it to all file locations within
+    the config file. Individual locations are put into the config.  Don't have
+    to put entire filepath location for each layer. Works by looping over
+    config dictionary and subdictionary to fine locations named 'file'.
+
+    Args:
+        input_path (str): Path that needs to be appended to the front of all
+            the file names/paths in config.
+        config (ConfigObj): Object defining the model and its inputs.
+
+    Returns:
+        config dictionary with complete file paths.
+
+    """
+
+    # Pull all other filepaths that need editing
+    for keys1 in config.keys():
+        outer = keys1
+        for keys2 in config[outer].keys():
+            second = keys2
+            if hasattr(config[outer][second], "keys") is False:
+                if second == "slopefile" or second == "file":
+                    path_to_correct = config[outer][second]
+                    config[outer][second] = os.path.join(input_path, path_to_correct)
+            else:
+                for keys3 in config[outer][second].keys():
+                    third = keys3
+                    if hasattr(config[outer][second][third], "keys") is False:
+                        if third == "file" or third == "filepath":
+                            path_to_correct = config[outer][second][third]
+                            config[outer][second][third] = os.path.join(
+                                input_path, path_to_correct
+                            )
+                    else:
+                        for keys4 in config[outer][second][third].keys():
+                            fourth = keys4
+                            if (
+                                hasattr(config[outer][second][third][fourth], "keys")
+                                is False
+                            ):
+                                if fourth == "file" or fourth == "filepath":
+                                    path_to_correct = config[outer][second][third][
+                                        fourth
+                                    ]
+                                    config[outer][second][third][fourth] = os.path.join(
+                                        input_path, path_to_correct
+                                    )
+                            else:
+                                for keys5 in config[outer][second][third][
+                                    fourth
+                                ].keys():
+                                    fifth = keys5
+                                    if fifth == "file" or fifth == "filepath":
+                                        path_to_correct = config[outer][second][third][
+                                            fourth
+                                        ][fifth]
+                                        config[outer][second][third][fourth][
+                                            fifth
+                                        ] = os.path.join(input_path, path_to_correct)
+
+    return config
